Fixed reducer for grading class of solutions
Still creates more copes than is necessary, but it now
only makes extra shallow copies of part of the state once
for each element in the current solution class. Rather
that the previous approach that would make a deep copy of
everything. Could consider trying to enhance the helper function,
to address this, but for now it should be okay.