Fix editing a step
This is a bit complicated, I currently am firing a change action for every keystroke.
It appears that when the MathInput component re-renders it moves the cursor to the end
of the box. The temporary fix was to remove the key from the top level container so that react
wouldn't know the step had changed, but this messes a bit with the render lifecycle and may cause issues
in the future.